requirements.txt
# Core GUI Framework
# PyQt5 installed via brew

# Computer Vision & Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
imagehash==4.3.1
exifread==3.0.0

# Deep Learning & AI
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.206  # YOLOv8
face-recognition==1.3.0
# dlib installed via dlib-bin

# OCR
pytesseract==0.3.10
easyocr==1.7.1

# Data Processing
numpy==1.24.3
pandas==2.1.3

# PDF Generation
reportlab==4.0.7
matplotlib==3.8.2

# Database
# sqlite3 is built-in to Python

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
loguru==0.7.2

# Build & Package
PyInstaller==6.2.0

# Development & Testing
pytest==7.4.3
pytest-qt==4.2.0
black==23.11.0
flake8==6.1.0